Move more shape events
